http오류처리

+++ src/memo/memo.service.ts
@@ -1,10 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException, HttpException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Memo } from './entity/memo.entity';
 import { MemoPostReqDto } from './dto/memo.post.request.dto';
 import { UpdateMemoDto } from './dto/memo.update.dto';
-import { throwIfEmpty } from 'rxjs';
 
 @Injectable()
 export class MemoService {
@@ -13,24 +12,47 @@ constructor(
     private memoRepository:Repository<Memo>,
 ){}
 
-async create(memoPostReqDto:MemoPostReqDto){
-    this.memoRepository.save(memoPostReqDto).then(()=>throwIfEmpty());
+async create(memoPostReqDto:MemoPostReqDto): Promise<Memo>{
+    const memo = this.memoRepository.create(memoPostReqDto);
+    return await this.memoRepository.save(memo);
 }
 
-findAll():Promise<Memo[]>{
-    return this.memoRepository.find();
+async findAll():Promise<Memo[]>{
+    const memoes = await this.memoRepository.find();
+    if(!memoes.length){
+        throw new NotFoundException('메모를 찾을 수 없습니다.');
+    }
+    return memoes;
 }
 
-findOne(id:number):Promise<Memo>{
-    return this.memoRepository.findOneBy({id});
+async findOne(id:number):Promise<Memo>{
+    const memo = await this.memoRepository.findOneBy({id});
+    if(!memo){
+        throw new NotFoundException(id+' 메모를 찾을 수 없습니다.');
+    }
+    return memo;
 }
 
 async update(id:number,updateMemoDto:UpdateMemoDto){
-    await this.memoRepository.update(id,updateMemoDto);
+    const memo = await this.memoRepository.findOneBy({id});
+    if(!memo){
+        throw new NotFoundException(id+' 메모를 찾을 수 없습니다.');
+    }
+    const result = await this.memoRepository.update(id,updateMemoDto);
+    if(result.affected===0){
+        throw new HttpException('메모 업데이트 실패',500);
+    }
 }
 
 async remove(id:number){
-    await this.memoRepository.delete(id);
+    const memo = await this.memoRepository.findOneBy({id});
+    if(!memo){
+        throw new NotFoundException(id+' 메모를 찾을 수 없습니다.');
+    }
+    const result = await this.memoRepository.delete(id);
+    if (result.affected===0) {
+        throw new HttpException('메모 삭제 실패', 500);
+    }
 }
